Created user message model
